Moving code around, adding a sidebar
