+++ nextjs-blog-starter-mdx/package.json
@@ -3,8 +3,8 @@
   "version": "1.0.1",
   "scripts": {
     "build": "next build && next export",
-    "demo-content": "demo-content --count=5 --dir=./_posts",
-    "images": "markdown-modifier --startdir=./_posts --destdir=./public/doc-images",
+    "demo-content": "demo-content --count=50 --dir=./_posts",
+    "images": "markdown-modifier --startdir=./_posts --destdir=doc-images --public=./public",
     "dev": "next",
     "saraswati:dev": "run-p dev images",
     "modifier": "markdown-modifier",
@@ -19,7 +19,7 @@
     "@next/mdx": "^10.0.3",
     "@saraswati-js/demo-content": "1.1.3",
     "@saraswati-js/markdown-finder": "^1.0.3",
-    "@saraswati-js/markdown-modifier": "^1.0.4",
+    "@saraswati-js/markdown-modifier": "1.0.5",
     "classnames": "2.2.6",
     "copy-paste": "^1.3.0",
     "date-fns": "2.10.0",
