+++ nextjs-blog-starter-mdx/src/design/components/sidebar/index.js
@@ -1,0 +1,29 @@
+import React from "react";
+import List from "@material-ui/core/List";
+import Divider from "@material-ui/core/Divider";
+
+import { oldContent as content } from '../../../_sample-data'
+import SidebarItem from './item'
+
+function Sidebar({ items = content, depthStep, depth, expanded }) {
+  return (
+    <div className="sidebar">
+      <List disablePadding dense>
+        {items.map((sidebarItem, index) => (
+          <React.Fragment key={`${sidebarItem.name}${index}`}>
+            {sidebarItem === "divider" ? (<Divider style={{ margin: "6px 0" }} />) : (
+              <SidebarItem
+                depthStep={depthStep}
+                depth={depth}
+                expanded={expanded}
+                item={sidebarItem}
+              />
+            )}
+          </React.Fragment>
+        ))}
+      </List>
+    </div>
+  );
+}
+
+export default Sidebar;
