+++ nextjs-blog-starter-mdx/src/design/pages/more-stories.js
@@ -1,5 +1,4 @@
 import Link from 'next/link'
-import Sidebar from '../components/sidebar'
 
 // import PostPreview from '../components/post-preview'
 // const isEmpty = obj => Object.keys(obj).length === 0
@@ -10,10 +9,6 @@ export default function MoreStories({ posts }) {
       <h2 className="mb-8 text-6xl md:text-7xl font-bold tracking-tighter leading-tight">
         More Stories
       </h2>
-      
-      <div className="left-side">
-        <Sidebar />
-      </div>
       
       <div className="body-side">
         {posts.map((post) => (
