+++ nextjs-blog-starter-mdx/src/styles/index.css
@@ -20,7 +20,7 @@
 }
 
 .left-side {
-  width: 600px;
+  /* width: 600px; */
   float: left;
 }
